refactor: task refs sorting, task constructors
add time-helper module, edit task Schema
